requirements.txt
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
atomicwrites==1.4.0
attrs==20.3.0
CD==1.1.0
certifi==2020.12.5
colorama==0.4.4
distlib==0.3.1
Django==3.1.4
django-cors-headers==3.7.0
django-filter==2.4.0
django-webapp==1.0.0
djangorestframework==3.12.2
filelock==3.0.12
iniconfig==1.1.1
isort==5.7.0
lazy-object-proxy==1.4.3
Markdown==3.3.3
mccabe==0.6.1
mysqlclient==2.0.3
packaging==20.8
Pillow==8.1.0
pipenv==2020.11.15
pluggy==0.13.1
protobuf==3.15.0
py==1.10.0
pylint==2.6.2
pylint-protobuf==0.17.0
PyMySQL==1.0.2
pyparsing==2.4.7
pytest==6.2.1
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
virtualenv==20.4.2
virtualenv-clone==0.5.4
wrapt==1.12.1
